fix: avoid cyclic dependency
